Update on f_runoff parameter -parametrization of RCN computation
f_runoff is the parameter associated with the computation of surface runoff using the RCN method. It is used to adjust the individual RCN values in the original HB version as follows:
RCNII * f_runoff
where RCNII represents the individual RCN values assigned to each cell of the modelling grid, ranging from 1 to 100. f_runoff is one of the eight parameters that must be calibrated by users.

Following the correction made to the RCN method in response to issue #30 , the parameter f_runoff is often >1, which introduces a model error by resulting in RCN values >100.
As a result, the parameterization of the RCNII correction was modified as follows::
100 - (100-RCNII) * 1/f_runoff