Further update on the RCN method
Still following issue #30 on the RCN method, the application of f_runoff to modify RCNII values needed an update because small RCN values and f_runoff<1 led to modified RCN values<0, thus producing <0 runoff.

The update ensures that no mater the value of f_runoff, modified RCN values will only range between 1 and 100.
This update impacts the acceptable range of values for f_runoff that is now >1. A test was performed in South-West Quebec and f_runoff optimized value was ~5. More updates to come on rechaRge-book.